requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
aiosqlite==0.20.0
apscheduler==3.10.4
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
